+++ src/button/button-test.jsx
@@ -37,6 +37,12 @@ describe('button', () => {
         expect(button.node).to.not.have.class('button_focused');
     });
 
+    it('should set "focused" class when focused=true', () => {
+        let button = render(<Button focused={ true } >Button-example</Button>);
+
+        expect(button.node).to.have.class('button_focused');
+    });
+
     it('should set/unset class on button hovered/unhovered', () => {
         let button = render(<Button>Button-example</Button>);
 
@@ -62,6 +68,21 @@ describe('button', () => {
         button.node.click();
 
         expect(onClick).to.have.been.called.once;
+    });
+
+    it('should unset class `hovered` and `focused` on disabled button', () => {
+        let button = render(<Button />);
+
+        simulate(button.node, 'mouseEnter');
+        button.instance.focus();
+
+        expect(button.node).to.have.class('button_hovered');
+        expect(button.node).to.have.class('button_focused');
+
+        button = render(<Button disabled={ true } />);
+
+        expect(button.node).to.not.have.class('button_hovered');
+        expect(button.node).to.not.have.class('button_focused');
     });
 
     it('should call `onFocus` callback after button was focused', (done) => {