+++ src/card-input/card-input.jsx
@@ -20,6 +20,7 @@ import performance from '../performance';
 @performance()
 class CardInput extends React.Component {
     static propTypes = {
+        ...Input.propTypes,
         /** Подсказка в текстовом поле */
         placeholder: Type.string
     };