+++ src/popup/popup.jsx
@@ -2,7 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { autobind, debounce } from 'core-decorators';
+import { autobind } from 'core-decorators';
+import debounce from 'lodash.debounce';
 import React from 'react';
 import Type from 'prop-types';
 
@@ -146,6 +147,12 @@ class Popup extends React.Component {
     popup;
     inner;
     content;
+
+    handleResizeWindow = debounce(() => {
+        if (this.isPropsToPositionCorrect()) {
+            this.redraw();
+        }
+    }, 200);
 
     componentWillMount() {
         if (this.context.isInCustomContainer
@@ -312,14 +319,6 @@ class Popup extends React.Component {
         this.redraw();
     }
 
-    @autobind
-    @debounce(200)
-    handleResizeWindow() {
-        if (this.isPropsToPositionCorrect()) {
-            this.redraw();
-        }
-    }
-
     /**
      * Задает элемент, к которому будет привязан popup.
      *
@@ -327,6 +326,10 @@ class Popup extends React.Component {
      * @param {HTMLElement} target Элемент, к которому будет привязан popup
      */
     setTarget(target) {
+        if (this.anchor === target) {
+            return;
+        }
+
         this.anchor = target;
         this.redraw();
     }