+++ CHANGELOG.md
@@ -1,3 +1,88 @@
+<a name="4.6.0"></a>
+# 4.6.0 (2017-05-26)
+
+
+### Bug Fixes
+
+* **button:** fix disabled button ([#36](https://github.com/alfa-laboratory/arui-feather/issues/36)) ([654b293](https://github.com/alfa-laboratory/arui-feather/commit/654b293))
+
+
+### Features
+
+* add accept prop for attach component ([#44](https://github.com/alfa-laboratory/arui-feather/issues/44)) ([992547d](https://github.com/alfa-laboratory/arui-feather/commit/992547d))
+
+
+
+<a name="4.5.0"></a>
+# 4.5.0 (2017-05-25)
+
+
+### Bug Fixes
+
+* fix prop types for composite components ([#40](https://github.com/alfa-laboratory/arui-feather/issues/40)) ([8dace92](https://github.com/alfa-laboratory/arui-feather/commit/8dace92))
+
+
+### Features
+
+* add defaultValue prop for textarea ([#39](https://github.com/alfa-laboratory/arui-feather/issues/39)) ([8aab55d](https://github.com/alfa-laboratory/arui-feather/commit/8aab55d))
+* docs for composed components ([#41](https://github.com/alfa-laboratory/arui-feather/issues/41)) ([318fcd4](https://github.com/alfa-laboratory/arui-feather/commit/318fcd4))
+
+
+
+<a name="4.4.0"></a>
+# 4.4.0 (2017-05-23)
+
+
+### Bug Fixes
+
+* fix sidebar viewport for android ([#37](https://github.com/alfa-laboratory/arui-feather/issues/37)) ([c2d884e](https://github.com/alfa-laboratory/arui-feather/commit/c2d884e))
+* **package.json:** add babel-runtime dependency ([c4472bc](https://github.com/alfa-laboratory/arui-feather/commit/c4472bc))
+* **playground:** fix url of playground ([#31](https://github.com/alfa-laboratory/arui-feather/issues/31)) ([5608be5](https://github.com/alfa-laboratory/arui-feather/commit/5608be5))
+* **popup:** replace deprecated debounce decorator ([#30](https://github.com/alfa-laboratory/arui-feather/issues/30)) ([5e1d640](https://github.com/alfa-laboratory/arui-feather/commit/5e1d640))
+
+
+### Features
+
+* add styleguide for input and for calendarInput ([#34](https://github.com/alfa-laboratory/arui-feather/issues/34)) ([44756f4](https://github.com/alfa-laboratory/arui-feather/commit/44756f4))
+* change react-textarea-autosize version ([#29](https://github.com/alfa-laboratory/arui-feather/issues/29)) ([07c8360](https://github.com/alfa-laboratory/arui-feather/commit/07c8360))
+* fix fonts for styleguide ([#33](https://github.com/alfa-laboratory/arui-feather/issues/33)) ([7067f46](https://github.com/alfa-laboratory/arui-feather/commit/7067f46))
+* **playground:** add share example functionality ([#23](https://github.com/alfa-laboratory/arui-feather/issues/23)) ([49cc2ed](https://github.com/alfa-laboratory/arui-feather/commit/49cc2ed))
+* **slide-down:** add component ([#27](https://github.com/alfa-laboratory/arui-feather/issues/27)) ([2e5d959](https://github.com/alfa-laboratory/arui-feather/commit/2e5d959))
+
+
+
+<a name="4.3.0"></a>
+# 4.3.0 (2017-05-11)
+
+
+### Bug Fixes
+
+* **form-field:** radio/checkbox group with view='line' in form-field ([#26](https://github.com/alfa-laboratory/arui-feather/issues/26)) ([c366e98](https://github.com/alfa-laboratory/arui-feather/commit/c366e98))
+* **input-autocomplete:** focus ([#28](https://github.com/alfa-laboratory/arui-feather/issues/28)) ([aeea353](https://github.com/alfa-laboratory/arui-feather/commit/aeea353))
+
+
+### Features
+
+* sidebar csstansition group animation ([#25](https://github.com/alfa-laboratory/arui-feather/issues/25)) ([4bbc349](https://github.com/alfa-laboratory/arui-feather/commit/4bbc349))
+* **demo:** add demo readme for several components ([#20](https://github.com/alfa-laboratory/arui-feather/issues/20)) ([bc1e221](https://github.com/alfa-laboratory/arui-feather/commit/bc1e221))
+
+
+
+<a name="4.2.1"></a>
+## 4.2.1 (2017-05-02)
+
+
+### Bug Fixes
+
+* **page:** use image instead of gradient, get rid of banding ([011d5fb](https://github.com/alfa-laboratory/arui-feather/commit/011d5fb))
+
+
+### Features
+
+* **demo:** add styleguidist demo ([#13](https://github.com/alfa-laboratory/arui-feather/issues/13)) ([1379f55](https://github.com/alfa-laboratory/arui-feather/commit/1379f55))
+
+
+
 <a name="4.2.0"></a>
 # 4.2.0 (2017-04-24)